useBabyの部分を削除し、ファイル名変更
